@@ -1,5 +1,7 @@
-package me.chatapp.stchat.network;
+package me.chatapp.stchat.api;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import me.chatapp.stchat.model.Message;
 import org.json.JSONObject;
 
 import java.io.BufferedReader;
@@ -7,11 +9,13 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.net.Socket;
+import java.util.function.Consumer;
 
 public class SocketClient {
     private final Socket socket;
     private final PrintWriter writer;
     private final BufferedReader reader;
+    private Consumer<Message> messageListener;
 
     public SocketClient(String host, int port) throws IOException {
         socket = new Socket(host, port);
@@ -42,13 +46,64 @@ public class SocketClient {
     public void close() {
         try {
             if (isConnected()) {
-                writer.println("/logout"); // gửi lệnh logout đến server
+                writer.println("/logout");
                 writer.flush();
                 socket.close();
             }
         } catch (IOException e) {
             System.err.println("Lỗi khi đóng kết nối Socket: " + e.getMessage());
         }
+    }
+
+    public void startListening(java.util.function.Consumer<String> messageConsumer) {
+        new Thread(() -> {
+            String line;
+            try {
+                while ((line = reader.readLine()) != null) {
+                    messageConsumer.accept(line);
+                }
+            } catch (IOException e) {
+                System.err.println("Disconnected from server.");
+            }
+        }).start();
+    }
+
+
+    public void setMessageListener(java.util.function.Consumer<Message> listener) {
+        this.messageListener = listener;
+    }
+
+    public void sendMessage(Message message) {
+        try {
+            ObjectMapper objectMapper = new ObjectMapper();
+            String json = objectMapper.writeValueAsString(message);
+            send(json);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    public void startListening() {
+        new Thread(() -> {
+            String line;
+            try {
+                ObjectMapper objectMapper = new ObjectMapper();
+                objectMapper.findAndRegisterModules(); // để hỗ trợ LocalDateTime
+
+                while ((line = reader.readLine()) != null) {
+                    try {
+                        Message message = objectMapper.readValue(line, Message.class);
+                        if (messageListener != null) {
+                            messageListener.accept(message);
+                        }
+                    } catch (Exception parseEx) {
+                        System.err.println("Không thể parse message: " + line);
+                    }
+                }
+            } catch (IOException e) {
+                System.err.println("Disconnected from server.");
+            }
+        }).start();
     }
 
     public void simulateLogin(String username, String password) throws IOException {
