@@ -18,7 +18,7 @@ import javafx.scene.text.FontWeight;
 import javafx.scene.text.Text;
 import javafx.stage.Stage;
 import javafx.util.Duration;
-import me.chatapp.stchat.network.SocketClient;
+import me.chatapp.stchat.api.SocketClient;
 import me.chatapp.stchat.util.AnimationUtil;
 import me.chatapp.stchat.util.DisplayUtil;
 import me.chatapp.stchat.util.ValidateUtil;
